feat: allow top level provider options
