+++ src/utils.ts
@@ -1,9 +1,12 @@
 import path from 'path'
 import util from 'util'
+import defu from 'defu'
 import consola from 'consola'
 import fetch from 'node-fetch'
 import hasha from 'hasha'
 import fs from 'fs-extra'
+import upath from 'upath'
+import { ModuleOptions, ModuleProvider } from 'types'
 
 const streamPipeline = util.promisify(require('stream').pipeline)
 
@@ -34,4 +37,86 @@ export function getFileExtension (url: string) {
 
 export function hash (value: string, length = 6) {
   return hasha(value).substr(0, length)
+}
+
+export async function getProviders (nuxt, options: ModuleOptions): Promise<ModuleProvider[]> {
+  const providerNames = await fs.readdir(path.resolve(__dirname, './providers'))
+
+  // enrich local provider
+  options.local = enrichLocalProvider(nuxt, options.local || {})
+
+  const providers = [
+    // official providers
+    ...providerNames.map(name => options[name] && [name, options[name]]).filter(Boolean),
+    // user custom providers
+    ...Object.entries(options.providers)
+  ]
+  return providers
+    .map(([key, provider]) => loadProvider(nuxt, key, provider))
+}
+
+export function enrichLocalProvider (nuxt, providerOptions) {
+  const { options } = nuxt
+  // Default port
+  const defaultPort =
+   process.env.PORT ||
+   process.env.npm_package_config_nuxt_port ||
+   (options.server && options.server.port) ||
+   3000
+
+  // Default host
+  let defaultHost =
+   process.env.HOST ||
+   process.env.npm_package_config_nuxt_host ||
+   (options.server && options.server.host) ||
+   'localhost'
+
+  /* istanbul ignore if */
+  if (defaultHost === '0.0.0.0') {
+    defaultHost = 'localhost'
+  }
+
+  // Default prefix
+  const prefix = '/'
+
+  providerOptions = defu(providerOptions, {
+    baseURL: `http://${defaultHost}:${defaultPort}${prefix}`,
+    dir: path.join('~', options.dir.static),
+    clearCache: false,
+    cacheDir: '~~/node_modules/.cache/nuxt-image',
+    sharp: {}
+  })
+
+  providerOptions.dir = nuxt.resolver.resolveAlias(providerOptions.dir)
+  providerOptions.cacheDir = nuxt.resolver.resolveAlias(providerOptions.cacheDir)
+
+  return providerOptions
+}
+
+export function loadProvider (nuxt, key: string, provider: any) {
+  if (typeof provider === 'string') {
+    provider = { provider }
+  } else if (typeof provider === 'object') {
+    provider = { options: provider }
+  }
+
+  if (!provider.name) {
+    provider.name = key
+  }
+  if (!provider.provider) {
+    provider.provider = provider.name
+  }
+  if (typeof provider.provider === 'string') {
+    provider.provider = tryRequire('./providers/' + provider.provider) ||
+      nuxt.resolver.requireModule(provider.provider)
+  }
+  const { middleware, runtime, runtimeOptions } = provider.provider(provider.options)
+  // TODO: verify provider.provider and warn+skip if invalid
+  return {
+    ...provider,
+    middleware,
+    runtimeOptions,
+    runtime: upath.normalize(runtime),
+    importName: 'runtime_' + hash(runtime).substr(0, 8)
+  }
 }
