+++ types/module.d.ts
@@ -22,3 +22,23 @@ export interface ModuleOptions {
 }
 
 export const imageModule: Module<ModuleOptions>
+
+export type ProviderServerMiddleware = (req, res, next) => void
+
+export type ProviderFactory = (options: any) => Provider
+
+export interface Provider {
+  runtime: string
+  runtimeOptions: any
+  middleware?(): ProviderServerMiddleware
+}
+
+export interface ModuleProvider {
+  name: string
+  importName: string
+  options: any
+  provider: ProviderFactory
+  runtime: string
+  runtimeOptions: any
+  middleware?: any
+}
